+++ gui/edit.js
@@ -28,7 +28,7 @@
  THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
  @author Johannes Heinecke
- @version 2.17.3 as of 25th September 2022
+ @version 2.17.4 as of 5th October 2022
  */
 
 
@@ -946,10 +946,137 @@ function isElementInViewport(el) {
 }
 
 
-
-
 // process shortcuts: we catch keys hit in the editor. If a word is active, we try to apply
+let timeout = null;
+
+var  shortcutseq = "";
+// Listen for keystroke events
 $(window).on('keypress', function (evt) {
+    if ($(".modal").is(":visible")) {
+        // if a model is open, we do not want to catch keypress events, since we are editing text
+        unsetPShC();
+        return;
+    }
+
+    else if (graphtype == 3) {
+        // in table mode, we need all keys to edit the table cells
+        unsetPShC();
+        return;
+    }
+
+    else if ($("#subtreesearch").is(":visible") && evt.which !== 33) {
+        // when editing a subtree, we need all keys to edit the table cells
+        unsetPShC();
+        return;
+    }
+
+    // hardwired keys
+    else if (evt.which == 63) { // "?"
+        unsetPShC();
+        ToggleShortcutHelp();
+        //} else if (evt.which == 33) { // "!"
+        //    ToggleSubtree();
+    } else if (evt.which == 43) { // "+"
+        unsetPShC();
+        sendmodifs({"cmd": "next"});
+    } else if (evt.which == 45) { // "-"
+        unsetPShC();
+        sendmodifs({"cmd": "prec"});
+    } else if (evt.which == 61) { // "=" validator
+        unsetPShC();
+        $("#valid").click();
+
+    } else if (evt.which == 38) { // "&" followed by three digits: put word with ID into viewport
+        // TODO
+        unsetPShC();
+        // move viewport to make token visible
+        token_to_be_focussed = 0;
+        $("#pendingshortcuts").text("&");
+        $("#psc").show();
+        return;
+    } else  if (evt.which == 95) { // "_" delete all features
+        sendmodifs({"cmd": "mod feat " + clickedNodes[0] + " " + "_"});
+        unsetPShC();
+        return;
+        
+    } else {
+
+        // Clear the timeout if it has already been set.
+        // This will prevent the previous task from executing
+        // if it has been less than <MILLISECONDS>
+        clearTimeout(timeout);
+        shortcutseq += String.fromCharCode(evt.which);
+        // Make a new timeout set to go off in 1000ms (1 second)
+        timeout = setTimeout(function () {
+            // process sequence
+            //    console.log('Input Value:', shortcutseq);  
+        
+            var newval = shortcutsUPOS[shortcutseq];
+            if (newval != undefined) {
+                //console.log("UPOS", newval);
+                sendmodifs({"cmd": "mod upos " + clickedNodes[0] + " " + newval});
+                shortcutseq = "";
+                deprels = [];
+                uposs = [];
+                clickedNodes = [];
+                unsetPShC();
+                return;
+            }
+
+            newval = shortcutsDEPL[shortcutseq];
+            if (newval != undefined) {
+                //console.log("DEPL", newval);
+                sendmodifs({"cmd": "mod deprel " + clickedNodes[0] + " " + newval});
+                shortcutseq = "";
+                deprels = [];
+                uposs = [];
+                clickedNodes = [];
+                unsetPShC();
+                return;
+            }
+
+            newval = shortcutsFEATS[shortcutseq];
+            if (newval != undefined) {
+                sendmodifs({"cmd": "mod addfeat " + clickedNodes[0] + " " + newval});
+                shortcutseq = "";
+                deprels = [];
+                uposs = [];
+                clickedNodes = [];
+                unsetPShC();
+                return;
+            }
+
+            newval = shortcutsXPOS[shortcutseq];
+            if (newval != undefined) {
+                //console.log("XPOS", newval, newval[0]);
+                if (newval.length > 1) {
+                    // change UPOS and XPOS
+                    sendmodifs({"cmd": "mod pos " + clickedNodes[0] + " " + newval[1] + " " + newval[0]});
+                } else {
+                    // change only XPOS
+                    sendmodifs({"cmd": "mod xpos " + clickedNodes[0] + " " + newval[0]});
+                }
+                
+                shortcutseq = "";
+                deprels = [];
+                uposs = [];
+                clickedNodes = [];
+                unsetPShC();
+                return;
+            }
+
+        
+    
+        
+            shortcutseq = "";
+        }, 500);
+    }
+});
+
+// OLD
+// process shortcuts: we catch keys hit in the editor. If a word is active, we try to apply
+$(window).on('keypressQQQ', function (evt) {
+    return; 
     //console.log("AEVT", evt.which, evt.keyCode, String.fromCharCode(evt.keyCode), clickedNodes);
     //console.log("kk", evt.which, unprocessedkeystrokes);