+++ CHANGES.md
@@ -1,5 +1,8 @@
 # Changes
 
+## Version 2.17.5
+* new option --shortcutTimeout to set the maximal time (milliseconds) between to keys of a shortcut sequence
+
 ## Version 2.17.4
 * shortcuts sequence must be taped with max 300ms between keys. This permits having short and long shortcuts (e.g. A, and AV)
 * bug correction in complex search: child(child(...))